Split selector definition and configured selector
