+++ lib/capybara/selector/filters/locator_filter.rb
@@ -10,8 +10,18 @@ module Capybara
           super(nil, nil, block, options)
         end
 
-        def matches?(node, value, context = nil)
-          super(node, nil, value, context)
+        def matches?(node, value, context = nil, exact:)
+          apply(node, value, true, context, exact: exact)
+        rescue Capybara::ElementNotFound
+          false
+        end
+
+      private
+
+        def apply(subject, value, skip_value, ctx, **options)
+          return skip_value if skip?(value)
+
+          filter_context(ctx).instance_exec(subject, value, **options, &@block)
         end
       end
     end
