+++ lib/capybara/spec/session/find_spec.rb
@@ -235,7 +235,7 @@ Capybara::SpecHelper.spec '#find' do
 
   it 'should accept an XPath instance' do
     @session.visit('/form')
-    @xpath = Capybara::Selector[:fillable_field].call('First Name')
+    @xpath = Capybara::Selector.new(:fillable_field, config: {}).call('First Name')
     expect(@xpath).to be_a(::XPath::Union)
     expect(@session.find(@xpath).value).to eq('John')
   end
