+++ lib/capybara/spec/session/first_spec.rb
@@ -24,7 +24,7 @@ Capybara::SpecHelper.spec '#first' do
 
   it 'should accept an XPath instance' do
     @session.visit('/form')
-    @xpath = Capybara::Selector[:fillable_field].call('First Name')
+    @xpath = Capybara::Selector.new(:fillable_field, config: {}).call('First Name')
     expect(@xpath).to be_a(::XPath::Union)
     expect(@session.first(@xpath).value).to eq('John')
   end
