+++ spec/selector_spec.rb
@@ -164,25 +164,28 @@ RSpec.describe Capybara do
 
     describe 'xpath' do
       it 'uses filter names passed in' do
-        selector = Capybara::Selector.new :test do
+        Capybara.add_selector :test do
           xpath(:something, :other) { |_locator| XPath.descendant }
         end
+        selector = Capybara::Selector.new :test, config: nil
 
         expect(selector.expression_filters.keys).to include(:something, :other)
       end
 
       it 'gets filter names from block if none passed to xpath method' do
-        selector = Capybara::Selector.new :test do
+        Capybara.add_selector :test do
           xpath { |_locator, valid3:, valid4: nil| "#{valid3} #{valid4}" }
         end
+        selector = Capybara::Selector.new :test, config: nil
 
         expect(selector.expression_filters.keys).to include(:valid3, :valid4)
       end
 
       it 'ignores block parameters if names passed in' do
-        selector = Capybara::Selector.new :test do
+        Capybara.add_selector :test do
           xpath(:valid1) { |_locator, valid3:, valid4: nil| "#{valid3} #{valid4}" }
         end
+        selector = Capybara::Selector.new :test, config: nil
 
         expect(selector.expression_filters.keys).to include(:valid1)
         expect(selector.expression_filters.keys).not_to include(:valid3, :valid4)
@@ -202,25 +205,28 @@ RSpec.describe Capybara do
       end
 
       it 'uses filter names passed in' do
-        selector = Capybara::Selector.new :text do
+        Capybara.add_selector :test do
           css(:name, :other_name) { |_locator| '' }
         end
+        selector = Capybara::Selector.new :test, config: nil
 
         expect(selector.expression_filters.keys).to include(:name, :other_name)
       end
 
       it 'gets filter names from block if none passed to css method' do
-        selector = Capybara::Selector.new :test do
+        Capybara.add_selector :test do
           css { |_locator, valid3:, valid4: nil| "#{valid3} #{valid4}" }
         end
+        selector = Capybara::Selector.new :test, config: nil
 
         expect(selector.expression_filters.keys).to include(:valid3, :valid4)
       end
 
       it 'ignores block parameters if names passed in' do
-        selector = Capybara::Selector.new :test do
+        Capybara.add_selector :test do
           css(:valid1) { |_locator, valid3:, valid4: nil| "#{valid3} #{valid4}" }
         end
+        selector = Capybara::Selector.new :test, config: nil
 
         expect(selector.expression_filters.keys).to include(:valid1)
         expect(selector.expression_filters.keys).not_to include(:valid3, :valid4)
